==> pathology/00_medical_reports.ipynb <==
from utils.text_preprocessor import TextPreprocessor

## Read the data

import json

text_file = "H&E_IHC_plus"

with open(f"./data/medical_reports/{text_file}.json") as f:
    medical_reports_original = json.load(f)

medical_reports_original

## Initialize the text preprocessor

medical_reports_processor = TextPreprocessor(medical_reports_original, num_servers=5, verbose=True, use_heuristics=True)

preprocessed_medical_reports = medical_reports_processor.main_preprocessing()

medical_reports_processor.print_random_reports(preprocessed_medical_reports, 50)

with open(f"./data/medical_reports/{text_file}_preprocessed.json", "w") as f:
    json.dump(preprocessed_medical_reports, f, indent=4)

# Load in the preprocessed medical reports

with open(f"./data/medical_reports/H&E_IHC_plus_preprocessed.json") as f: 
    medical_reports = json.load(f)

medical_report_nrs = [
    report_nr
    for patient_reports in medical_reports.values()
    for report_nr in patient_reports.keys()
]
print(len(medical_report_nrs))

# Open all data
with open(f"./data/medical_reports/all_data.json") as f:
    all_data = json.load(f)

# Filter the medical report nrs, all_data has same structure as medical_reports
all_data_filtered = {
    patient_nr: {
        report_nr: report
        for report_nr, report in reports.items()
        if report_nr in medical_report_nrs
    }
    for patient_nr, reports in all_data.items()
}

# Filter empty patients and join the report contents with "' '.join(report.values())"
all_data_filtered = {
    patient_nr: {
        report_nr: ' '.join(report_content)
        for report_nr, report_content in report.items()
    }
    for patient_nr, report in all_data_filtered.items()
    if report
}

new_medical_report_nrs = [
    report_nr
    for patient_reports in all_data_filtered.values()
    for report_nr in patient_reports.keys()
]
print(len(new_medical_report_nrs))

with open(f"./data/medical_reports/H&E_IHC_plus_report_all_data_preprocessed.json", "w") as f:
    json.dump(all_data_filtered, f, indent=4)
